authorisation added

+++ handlers/adminhandler.py
@@ -1,4 +1,4 @@
-from tornado.web import RequestHandler
+from tornado.web import RequestHandler, authenticated
 from tornado import gen
 
 class AdminHandler(RequestHandler):
@@ -7,20 +7,22 @@ class AdminHandler(RequestHandler):
 
     @gen.coroutine
     def get(self):
-        news = yield self.db.get_all_news()
-        self.render("news-admin.html", news=news)
+        if self.get_secure_cookie("admin"):
+            news = yield self.db.get_all_news()
+            self.render("news-admin.html", news=news)
+        else:
+            self.redirect("/login")
     
     @gen.coroutine
     def post(self):
-        if 'title' in self.request.arguments:
-            title = self.get_argument('title', strip=True) 
-            body = self.get_argument('body', strip=True)
-            news = yield self.db.add_news(title, body)
-            print(title, body)
-        elif 'id' in self.request.arguments:
-            id = self.get_argument('id', strip=True)
-            news = yield self.db.delete_news(id)
-            print(id)
-
-        self.render("news-admin.html", news=news)
-        
+        if self.get_secure_cookie("admin"):
+            if 'title' in self.request.arguments:
+                title = self.get_argument('title', strip=True) 
+                body = self.get_argument('body', strip=True)
+                yield self.db.add_news(title, body)
+            elif 'id' in self.request.arguments:
+                id = self.get_argument('id', strip=True)
+                yield self.db.delete_news(id)
+            self.redirect("/admin")
+        else:
+            self.redirect("/login")
